refactor: common 컴포넌트 역할별 디렉토리 구조 세분화
