+++ components/auth/AdminProtectedRoute.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import { clearAuthCookie } from "@/lib/server/cookies";
 import axios from "axios";
-import LoadingSpinner from "@/components/common/LoadingSpinner";
+import LoadingSpinner from "@/components/common/loading/LoadingSpinner";
 
 export default function AdminProtectedRoute({ children }) {
   const router = useRouter();
