+++ components/common/reports/ReportDetailLayout.jsx
