+++ components/common/reports/ReportTableLayout.jsx
@@ -1,4 +1,4 @@
-import EmptyState from "@/components/common/EmptyState";
+import EmptyState from "@/components/common/feedback/EmptyState";
 import React from "react";
 
 function ReportTableLayout({ icon, emptyMessage, columns, rows }) {
