+++ components/dashboard/UserStatusDoughnutChart.jsx
@@ -9,7 +9,7 @@ import {
   Tooltip,
   ResponsiveContainer,
 } from "recharts";
-import EmptyState from "@/components/common/EmptyState";
+import EmptyState from "@/components/common/feedback/EmptyState";
 import {
   chartContainerBox,
   chartEmptyBox,
