+++ components/reports/review/AdminReviewReportsContainer.jsx
@@ -4,7 +4,7 @@ import { useCallback, useState } from "react";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { getAdminReports } from "@/services/adminReports";
 import { useAdminAuth } from "@/context/auth/useAdminAuth";
-import LoadingSpinner from "@/components/common/LoadingSpinner";
+import LoadingSpinner from "@/components/common/loading/LoadingSpinner";
 import ReviewReportView from "@/components/reports/review/ReviewReportView";
 import { ADMIN_REPORTS } from "@/constants/queryKeys";
